null safety and bring all up to date

+++ api/bin/server.dart
@@ -2,15 +2,29 @@ import 'package:arrahma_shared/shared.dart';
 import 'package:arrahma_web_api/api.dart';
 import 'package:arrahma_web_api/services/data_sync_service.dart';
 
+import 'package:conduit_core/conduit_core.dart';
+
 Future main() async {
   final isDebugging = Platform.environment['DEBUG'] == 'true';
+  final certPath = Platform.environment['CERT_PATH'];
+  final keyPath = Platform.environment['KEY_PATH'];
+  final portVal = Platform.environment['PORT'];
+  final port = (portVal != null ? int.tryParse(portVal) : null) ?? 8888;
+
+  if (isDebugging) {
+    print('Debugging mode enabled.');
+  }
+
+  if (!isDebugging && (certPath == null || keyPath == null)) {
+    print('Certificate and key paths are required.');
+    return;
+  }
   final app = Application<ArrahmaChannel>()
     ..isolateStartupTimeout = Duration(minutes: 30 * (isDebugging ? 4 : 1))
     ..options.configurationFilePath = 'config.yaml'
-    ..options.certificateFilePath = Platform.environment['CERT_PATH']
-    ..options.privateKeyFilePath = Platform.environment['KEY_PATH']
-    ..options.port =
-        int.tryParse(Platform.environment['PORT'] ?? 8888.toString());
+    ..options.certificateFilePath = certPath
+    ..options.privateKeyFilePath = keyPath
+    ..options.port = port;
 
   await FileService().delete(DataSyncService.idFilePath);
